backend/requirements.txt
Django>=4.0
djangorestframework
djangorestframework-simplejwt
gunicorn
psycopg2-binary
bs4
requests